Keep other mappings functional when W/E/K/J are not mapped.
Instead of using :map to W for the duplicate <C-Right> and L mappings, duplicate the actual call with :noremap so that they are kept functional when the W key is not mapped (:let g:csv_nomap_w = 1). I personally don't need three alternatives for the same movement and prefer to keep the built-in W motion.

Also, the <SID>MoveOver() function should not rely on the E key mapping.