movement: fix movement mappings J/K
Previously, the plugin tried to keep the cursor in the same spot within
a cell, if it moved upwards/downwards. This obviously failed, if the
width of the destination cell was different than of the starting cell.

So only do that, if the width of both cells is the same.

Also document, that this is a best effort approach here.

Fixes #139