Fix ArrangeColumn for vartabs
When using varibale tabstops, the first time a file is opened, the
column width are calculated correctly. Howver, all subsequent
calculations were wrong, leading to increasingly wide separation.

To repro:

* vim -c 'set ts=2 noet' test.tsv
* A<Tab>BCD<Tab>E
* AAAA<Tab>B<Tab>E
* :wq
* vim -c 'set ts=2 noet' test.tsv
* observe columns aligned, but too wide
* GoAAAA----------<Tab>B<Tab>E
* :%CSVArrangeColumn
* observe columns aligned, but too wide
* GoAAAA----------==========<Tab>B<Tab>E
* :%CSVArrangeColumn
* Observe that the BCD column is 10 chars too wide

With the fix, it's the right width in all 3 places, based on the tabstop
value of `2`

The calulation of column widths uses strdisplaywidth(), which itself is
dependent on the current value of the tabstops. To avoid this, we
temporarily disable variable tabstops while calculating the column
widths.