Fix: Incorrect parsing of custom /format/ when performed on a range.
When running a command on a range of rows, like   :'<,'>SumCol 1 /.:,/
the statements that parse the /format/ argument were reporting that it
had been entered with the wrong syntax.

The variable `start` should not have been used in the `matchstr`
statement. It has a different purpose.

:SumRow /.:,/ was failing because `start` was an undefined variable. The
extra parameters for matchstr were not needed here either.